Build pilot dataset
Adopt the data from CaCO3_NWP project to fit the requirement of
BERT-like model. The raw spectra data is in data/raw folder.
The compiled and masked spectral files (in csv) are in data/spe and
data/masked folders, respectively. Also, the mask files are exported to
data/mask folder.

	modified:   .gitignore
	new file:   pilot_01.ipynb