Modify the pretrain scripts from MAE for our work
However, after series of try and error, I think there are still too many
errors to be fixed. Since the codes are very dedicated and complicated,
it's not easy to debug during our modification. So I decide to build
from scratch. Step by step to see the errors and fix them.

	new file:   engine_pretrain.py
	new file:   main_pretrain.py
	new file:   util/lr_sched.py
	new file:   util/misc.py